[change] 古いタイプのリアクションデータをModel層に切り出しました
