+++ model/src/main/java/net/pantasystem/milktea/model/notes/reaction/LegacyReaction.kt
@@ -1,0 +1,21 @@
+package net.pantasystem.milktea.model.notes.reaction
+
+object LegacyReaction {
+
+    val reactionMap = mapOf(
+        "angry" to "\uD83D\uDCA2",  // 💢
+        "confused" to "\uD83D\uDE22", // 😢
+        "congrats" to "\uD83C\uDF89", // 🎉
+        "hmm" to "\uD83E\uDD14",    //  🤔
+        "laugh" to "\uD83D\uDE06",  // 😆
+        "like" to "\uD83D\uDC4D", // 👍
+        "love" to "❤", //❤
+        "pudding" to "\uD83C\uDF6E", // 🍮
+        "rip" to "\uD83D\uDE07", // 😇
+        "surprise" to "\uD83D\uDE2E", // 😮
+        "star" to "⭐" // ⭐
+    )
+
+    val defaultReaction = listOf("angry", "confused", "congrats", "hmm", "laugh", "like", "love", "pudding", "rip", "surprise", "star")
+
+}
